ref: Separate cadence from metrics feature
