+++ sentry-core/Cargo.toml
@@ -26,7 +26,8 @@ client = ["rand"]
 # and macros actually expand features (and extern crate) where they are used!
 debug-logs = ["dep:log"]
 test = ["client"]
-UNSTABLE_metrics = ["dep:cadence", "sentry-types/UNSTABLE_metrics"]
+UNSTABLE_metrics = ["sentry-types/UNSTABLE_metrics"]
+UNSTABLE_cadence = ["dep:cadence", "UNSTABLE_metrics"]
 
 [dependencies]
 cadence = { version = "0.29.0", optional = true }
